Fixed #19: Support command arguments
By defining `make_parser()` in a command that returns an instance of argparse.ArgumentParser,
arguments will be automatically parsed using that when entering commands in the REPL. The action has
been changed to take the optional args argument, which is the result of
`make_parser().parse_args(rest.split())`, where the rest means everything after the command name.

The default is to not have a parser defined, which means that `action(None)` is called.

Example of a command supporting arguments:

```
from .command import Command
from argparse import ArgumentParser

class TestCommand(Command):
  def name(self):
      return "test"

  def description(self):
      return "test command."

  def aliases(self):
      return ["t"]

  def make_parser(self):
    parser = ArgumentParser(prog = self.name())
    parser.add_argument("--wow")
    parser.add_argument('integers', metavar='int', type=int, choices=range(10),
                        nargs='+', help='an integer in the range 0..9')
    return parser

  def action(self, args = None):
      print("Test args={}".format(args))
```