Try parsing no arguments so default values are returned
This is ignored for commands that doesn't have an argument parser defined. But for others it tries
parsing even with no arguments so default values are returned in the arguments namespace, if
any. Before it would just give `args = None` instead of `Namespace(count=100)`, for instance.

Fixed log and workspace commands to conform.